Moves errors into a separate file error.rs
